Rename module io.tidy -> io.seglst
