+++ meeteval/io/seglst.py
